Change ErrorModel Code data type to short

+++ src/Core/VSGBulgariaMarketplace.Application/Models/Exceptions/ErrorModel.cs
@@ -2,7 +2,7 @@
 {
     internal class ErrorModel
     {
-        public int Code { get; set; }
+        public short Code { get; set; }
 
         public string ErrorMessage { get; set; }
     }
